services/ai/requirements.txt
fastapi==0.115.12
uvicorn[standard]==0.34.2
python-multipart==0.0.20
python-dotenv==1.1.0
pydantic==2.11.3
pydantic-settings==2.9.1

langchain>=0.3.25
langchain-core>=0.3.58
langchain-groq>=0.3.2
langgraph>=0.4.7

qdrant-client>=1.14.0

sentence-transformers>=4.0.0

pypdf>=5.0.0
python-docx>=1.1.0

httpx>=0.28.0
